Crate eventually_core

Container for the fundamental types and abstraction to support Event Sourcing and Command-Query Responsibility Segregation patterns.

Note

Generally, you should not depend from this crate directly, but instead import the public crate: eventually.

Please refer to the public crate API documentation for more information.

Modules

aggregate

Foundation traits for creating Domain abstractions using the Aggregate pattern.

projection

Contain support for Projection, an optimized read model of an Aggregate or of a number of Aggregates.

repository

Contains the Repository pattern implementation for AggregateRoot instances.

store

Contains the Event Store trait for storing and streaming Aggregate::Events.

subscription

Module for creating and managing long-running Subscriptions to incoming events in the EventStore.

versioning

Contains support for Optimistic Concurrency Control through Versioning.