Module eventually_core::aggregate

Foundation traits for creating Domain abstractions using the Aggregate pattern.

For ease of development, the following modules contain different traits for modeling Aggregate-compatible structures:

referential, containing an Aggregate-compatible trait where the trait is implemented for self, thus making State as Self.

Traits

Aggregate
An Aggregate is an entity which State is composed of one or more value-objects, entities or nested aggregates.

AggregateExt
Extension trait for Aggregate containing combinator functions.

Identifiable

Type Definitions

ErrorOf
Alias for the Error type of an Aggregate.

EventOf
Alias for the Event type of an Aggregate.

StateOf
Alias for the State type of an Aggregate.