Trait eventually_core::aggregate::Aggregate

pub trait Aggregate {
    type State;
    type Event;
    type Error;
    fn apply(
        state: Self::State, 
        event: Self::Event
    ) -> Result<Self::State, Self::Error>;
}

An Aggregate is an entity which State is composed of one or more value-objects, entities or nested aggregates.

State mutations are expressed through clear Domain Events which, if applied in the same order as they happened chronologically, will yield the same State value.

Associated Types

type State

State of the Aggregate.

Usually this associate type is either Self, Option<Self> or Option<T>, depending on whether the Aggregate state is defined in a separate data structure or using the same structure that implements this trait.

type Event

Domain events that express mutations of the Aggregate's State.

Usually, this type is an enum containing all possible Domain Events possible for this Aggregate.

type Error

Error type returned in apply when mutating the Aggregate State to the next version fails.

Usually, this error is a validation error type raised when the domain event that is being applied is invalid, based on the current State.

Consider using std::convert::Infallible (or ! type if using nightly) if the apply method doesn't fail.

Loading content...
Required methods

fn apply(
state: Self::State,
event: Self::Event
) -> Result<Self::State, Self::Error>

Applies the Event to the current State, returning either the next State or an Error.

Loading content...