Module eventually_core::subscription

Module for creating and managing long-running Subscriptions to incoming events in the EventStore.

Structs

Transient
Subscription type which gets deleted once the process using it gets terminated.

Enums

Error
Error type returned by a Transient Subscription.

Traits

EventSubscriber
Component to let users subscribe to newly-inserted events into the EventStore.

Subscription
A Subscription to an EventStream which can be "checkpointed": keeps a record of the latest message processed by itself using checkpoint, and can resume working from such message by using the resume.

Type Definitions

EventStream
Stream of events returned by the EventSubscriber::subscribe_all method.

SubscriptionStream
Stream of events returned by the Subscription::resume method.