Module eventually_core::aggregate

Foundation traits for creating Domain abstractions using the Aggregate pattern.

Structs

AggregateRoot
An AggregateRoot represents an handler to the Aggregate it's managing, such as:

AggregateRootBuilder
Builder type for new AggregateRoot instances.

Traits

Aggregate
An Aggregate manages a domain entity State, acting as a transaction boundary.

AggregateExt
Extension trait with some handy methods to use with Aggregates.

Type Definitions

AggregateId
A short extractor type for the Aggregate Id.