Struct eventually_core::subscription::Transient

pub struct Transient<Store, Subscriber> { /* fields omitted */ }

Subscription type which gets deleted once the process using it gets terminated.

Useful for in-memory or one-off Projections.

Implementations

impl<Store, Subscriber> Transient<Store, Subscriber>

pub fn new(store: Store, subscriber: Subscriber) -> Self

Creates a new Subscription using the specified EventStore and EventSubscriber to create the SubscriptionStream from.

pub fn from(self, sequence_number: u32) -> Self

Specifies the sequence number of the Event the SubscriptionStream should start from when calling run.