Trait eventually_core::subscription::Subscription

pub trait Subscription {
    type SourceId: Eq;
    type Event;
    type Error;
    fn resume(&self) -> BoxFuture<Result<SubscriptionStream<Self>, Self::Error>>;

    fn checkpoint(&self, version: u32) -> BoxFuture<Result<(), Self::Error>>;
}

A Subscription to an [EventStream] which can be "checkpointed": keeps a record of the latest message processed by itself using checkpoint, and can resume working from such message by using the resume.

Associated Types

type SourceId: Eq

Type of the Source id, typically an AggregateId.

type Event

Event type stored in the EventStore, typically an Aggregate::Event.

type Error

Possible errors returned when receiving events from the notification channel.

Loading content...
Required methods

fn resume(&self) -> BoxFuture<Result<SubscriptionStream<Self>, Self::Error>>

Resumes the current state of a Subscription by returning the EventStream, starting from the last event processed by the Subscription.

fn checkpoint(&self, version: u32) -> BoxFuture<Result<(), Self::Error>>

Saves the provided version (or sequence number) as the latest version processed.

Loading content...